# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Rede Social",
    "description": "API REST para uma rede social com postagens, comentários, curtidas e autenticação JWT.",
    "version": "1.0.0",
    "contact": {
      "name": "Sara Amabili Castilhos"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Insira o token JWT no formato: Bearer <token>"
    }
  },
  "tags": [
    { "name": "Autenticação", "description": "Registro e login de usuários" },
    { "name": "Postagens", "description": "CRUD de postagens (requer token Bearer)" },
    { "name": "Curtidas", "description": "Curtir e descurtir postagens (requer token Bearer)" },
    { "name": "Comentários", "description": "Gerenciar comentários em postagens (requer token Bearer)" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Registrar um novo usuário",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/RegisterBody" }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "schema": { "$ref": "#/definitions/AuthResponse" }
          },
          "400": {
            "description": "E-mail já cadastrado ou dados inválidos",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": { "description": "Erro interno do servidor" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Login de usuário existente",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/LoginBody" }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado – retorna token JWT",
            "schema": { "$ref": "#/definitions/AuthResponse" }
          },
          "401": {
            "description": "Credenciais inválidas",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": { "description": "Erro interno do servidor" }
        }
      }
    },
    "/api/posts": {
      "get": {
        "tags": ["Postagens"],
        "summary": "Listar todas as postagens",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de postagens com autores e comentários",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/Post" } }
          },
          "401": { "description": "Token não fornecido ou inválido" },
          "500": { "description": "Erro interno do servidor" }
        }
      },
      "post": {
        "tags": ["Postagens"],
        "summary": "Criar uma nova postagem",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/PostBody" }
          }
        ],
        "responses": {
          "201": {
            "description": "Postagem criada com sucesso",
            "schema": { "$ref": "#/definitions/Post" }
          },
          "400": {
            "description": "Dados inválidos",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "401": { "description": "Token não fornecido ou inválido" }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "tags": ["Postagens"],
        "summary": "Buscar uma postagem por ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID da postagem (MongoDB ObjectId)"
          }
        ],
        "responses": {
          "200": {
            "description": "Postagem encontrada",
            "schema": { "$ref": "#/definitions/Post" }
          },
          "401": { "description": "Token não fornecido ou inválido" },
          "404": { "description": "Post não encontrado" },
          "500": { "description": "Erro interno do servidor" }
        }
      },
      "put": {
        "tags": ["Postagens"],
        "summary": "Atualizar uma postagem (somente o autor)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID da postagem"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/PostBody" }
          }
        ],
        "responses": {
          "200": {
            "description": "Postagem atualizada com sucesso",
            "schema": { "$ref": "#/definitions/Post" }
          },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Token não fornecido ou inválido" },
          "403": { "description": "Você só pode editar seus próprios posts" },
          "404": { "description": "Post não encontrado" }
        }
      },
      "delete": {
        "tags": ["Postagens"],
        "summary": "Remover uma postagem (somente o autor)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID da postagem"
          }
        ],
        "responses": {
          "200": {
            "description": "Post removido com sucesso",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          },
          "401": { "description": "Token não fornecido ou inválido" },
          "403": { "description": "Você só pode deletar seus próprios posts" },
          "404": { "description": "Post não encontrado" },
          "500": { "description": "Erro interno do servidor" }
        }
      }
    },
    "/api/posts/{id}/like": {
      "post": {
        "tags": ["Curtidas"],
        "summary": "Curtir ou descurtir uma postagem",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID da postagem"
          }
        ],
        "responses": {
          "200": {
            "description": "Operação de curtida realizada",
            "schema": { "$ref": "#/definitions/LikeResponse" }
          },
          "401": { "description": "Token não fornecido ou inválido" },
          "404": { "description": "Post não encontrado" },
          "500": { "description": "Erro interno do servidor" }
        }
      }
    },
    "/api/posts/{id}/comments": {
      "post": {
        "tags": ["Comentários"],
        "summary": "Adicionar um comentário a uma postagem",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID da postagem"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CommentBody" }
          }
        ],
        "responses": {
          "201": {
            "description": "Comentário adicionado com sucesso",
            "schema": { "$ref": "#/definitions/Comment" }
          },
          "400": { "description": "Dados inválidos (comentário vazio)" },
          "401": { "description": "Token não fornecido ou inválido" },
          "404": { "description": "Post não encontrado" }
        }
      }
    },
    "/api/posts/{postId}/comments/{commentId}": {
      "delete": {
        "tags": ["Comentários"],
        "summary": "Remover um comentário (autor do comentário ou dono do post)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "type": "string",
            "description": "ID da postagem"
          },
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "type": "string",
            "description": "ID do comentário"
          }
        ],
        "responses": {
          "200": {
            "description": "Comentário removido",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          },
          "401": { "description": "Token não fornecido ou inválido" },
          "403": { "description": "Não autorizado" },
          "404": { "description": "Post ou comentário não encontrado" },
          "500": { "description": "Erro interno do servidor" }
        }
      }
    }
  },
  "definitions": {
    "RegisterBody": {
      "type": "object",
      "required": ["name", "email", "password"],
      "properties": {
        "name": { "type": "string", "example": "Sara Castilhos" },
        "email": { "type": "string", "example": "sara@email.com" },
        "password": { "type": "string", "example": "senha123" }
      }
    },
    "LoginBody": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": { "type": "string", "example": "sara@email.com" },
        "password": { "type": "string", "example": "senha123" }
      }
    },
    "AuthResponse": {
      "type": "object",
      "properties": {
        "_id": { "type": "string", "example": "64f1a2b3c4d5e6f7a8b9c0d1" },
        "name": { "type": "string", "example": "Sara Castilhos" },
        "email": { "type": "string", "example": "sara@email.com" },
        "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
      }
    },
    "PostBody": {
      "type": "object",
      "required": ["title", "content"],
      "properties": {
        "title": { "type": "string", "example": "Meu primeiro post" },
        "content": { "type": "string", "example": "Conteúdo do post aqui." },
        "imageUrl": { "type": "string", "example": "https://exemplo.com/imagem.jpg" }
      }
    },
    "Post": {
      "type": "object",
      "properties": {
        "_id": { "type": "string", "example": "64f1a2b3c4d5e6f7a8b9c0d2" },
        "title": { "type": "string", "example": "Meu primeiro post" },
        "content": { "type": "string", "example": "Conteúdo do post aqui." },
        "imageUrl": { "type": "string", "example": "https://exemplo.com/imagem.jpg" },
        "author": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "name": { "type": "string" },
            "email": { "type": "string" }
          }
        },
        "likes": { "type": "array", "items": { "type": "string" } },
        "comments": { "type": "array", "items": { "$ref": "#/definitions/Comment" } },
        "createdAt": { "type": "string", "example": "2024-01-01T00:00:00.000Z" },
        "updatedAt": { "type": "string", "example": "2024-01-01T00:00:00.000Z" }
      }
    },
    "CommentBody": {
      "type": "object",
      "required": ["text"],
      "properties": {
        "text": { "type": "string", "example": "Que post incrível!" }
      }
    },
    "Comment": {
      "type": "object",
      "properties": {
        "_id": { "type": "string", "example": "64f1a2b3c4d5e6f7a8b9c0d3" },
        "text": { "type": "string", "example": "Que post incrível!" },
        "author": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "name": { "type": "string" },
            "email": { "type": "string" }
          }
        },
        "createdAt": { "type": "string", "example": "2024-01-01T00:00:00.000Z" }
      }
    },
    "LikeResponse": {
      "type": "object",
      "properties": {
        "likesCount": { "type": "integer", "example": 5 },
        "liked": { "type": "boolean", "example": true }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": { "type": "string", "example": "Mensagem de erro" }
      }
    },
    "MessageResponse": {
      "type": "object",
      "properties": {
        "msg": { "type": "string", "example": "Operação realizada com sucesso" }
      }
    }
  }
}